+++ Changelog.md
@@ -2,7 +2,7 @@
 2.1.3.9 Beta (not yet available)
 ============
 
-
+- Optional mode added to use AviSynth/VapourSynth via VFW instead of directly.
 
 
 2.1.3.8 Beta